feat: 블로그 이동시 SafariController 사용
