--- burstcamp/burstcamp/Coordinator/App/AppCoordinatorEvent.swift
@@ -1,13 +1,0 @@
-//
-//  CoordinatorEvent.swift
-//  Eoljuga
-//
-//  Created by youtak on 2022/11/16.
-//
-
-import Foundation
-
-enum AppCoordinatorEvent {
-    case moveToAuthFlow
-    case moveToTabBarFlow
-}
