--- burstcamp/burstcamp/Coordinator/Auth/AuthCoordinatorEvent.swift
@@ -1,16 +1,0 @@
-//
-//  AuthCoordinatorEvent.swift
-//  Eoljuga
-//
-//  Created by neuli on 2022/11/22.
-//
-
-import Foundation
-
-enum AuthCoordinatorEvent {
-    case moveToDomainScreen
-    case moveToIDScreen
-    case moveToBlogScreen
-    case moveToTabBarScreen
-    case showAlert(String)
-}
