--- burstcamp/burstcamp/Coordinator/Home/HomeCoordinatorEvent.swift
@@ -1,12 +1,0 @@
-//
-//  HomeCoordinatorEvent.swift
-//  burstcamp
-//
-//  Created by youtak on 2022/12/07.
-//
-
-import Foundation
-
-enum HomeCoordinatorEvent {
-    case moveToFeedDetail(feed: Feed)
-}
