+++ burstcamp/burstcamp/Coordinator/ScrapPage/ScrapPageCoordinator.swift
@@ -6,6 +6,7 @@
 //
 
 import Combine
+import SafariServices.SFSafariViewController
 import UIKit
 
 protocol ScrapPageCoordinatorProtocol: TabBarChildCoordinator {
@@ -42,8 +43,24 @@ extension ScrapPageCoordinator {
 
     func moveToFeedDetail(feed: Feed) {
         let feedDetailViewController = prepareFeedDetailViewController(feed: feed)
-
+        sinkFeedViewController(feedDetailViewController)
         self.navigationController.pushViewController(feedDetailViewController, animated: true)
+    }
+
+    func moveToBlogSafari(url: URL) {
+        let safariViewController = SFSafariViewController(url: url)
+        self.navigationController.present(safariViewController, animated: true)
+    }
+
+    private func sinkFeedViewController(_ feedDetailViewController: FeedDetailViewController) {
+        feedDetailViewController.coordinatorPublisher
+            .sink { [weak self] event in
+                switch event {
+                case .moveToBlogSafari(let url):
+                    self?.moveToBlogSafari(url: url)
+                }
+            }
+            .store(in: &cancelBag)
     }
 
     private func prepareFeedDetailViewController(feed: Feed) -> FeedDetailViewController {
