--- burstcamp/burstcamp/Coordinator/ScrapPage/ScrapPageCoordinatorEvent.swift
@@ -1,12 +1,0 @@
-//
-//  ScrapPageCoordinatorEvent.swift
-//  burstcamp
-//
-//  Created by youtak on 2022/12/07.
-//
-
-import Foundation
-
-enum ScrapPageCoordinatorEvent {
-    case moveToFeedDetail(feed: Feed)
-}
