--- burstcamp/burstcamp/Coordinator/TabBar/TabBarCoordinatorEvent.swift
@@ -1,12 +1,0 @@
-//
-//  TabBarCoordinatorEvent.swift
-//  Eoljuga
-//
-//  Created by neuli on 2022/11/22.
-//
-
-import Foundation
-
-enum TabBarCoordinatorEvent {
-    case moveToAuthFlow
-}
